Allow preupload hooks to ignore more formatting ktlint rules
This CL adds a --no-verify-format rule to ktlint.py so that preuploads
can ignore more formatting rules. This will be used by projects who
already use ktfmt for formatting new files, so that useful ktlint errors
(like unused variables) are not hidden by a ton of ktlint formatting
issues which we are now mostly ignoring.

Bug: 235461679
Test: Manual
Change-Id: I324bcaf4e9e2b087201810343471a55f2434339f